librarify CYP
