+++ cyp.cabal
@@ -3,13 +3,23 @@ version:             0.1.0.0
 build-type:          Simple
 cabal-version:       >=1.16
 
+library
+  default-language:  Haskell2010
+  hs-source-dirs:    src
+  ghc-options:       -W
+  exposed-modules:   Test.Info2.Cyp
+  build-depends:
+      base ==4.*
+    , haskell-src-exts ==1.13.5.*
+    , parsec ==3.1.*
+    , pretty ==1.1.*
+    , pretty-show ==1.6.*
+
 executable cyp
   main-is:           Main.hs
   default-language:  Haskell2010
   ghc-options:       -W
   build-depends:
-    base ==4.*,
-    parsec ==3.1.*,
-    pretty ==1.1.*,
-    pretty-show ==1.6.*,
-    haskell-src-exts ==1.13.5.*
+      base ==4.*
+    , cyp
+    , pretty ==1.1.*
