+++ examples/length-append.cthy
