+++ examples/trivial.cprf
