+++ examples/trivial.cthy
